requirements.txt
Flask==2.0.2
PyJWT==2.3.0
pymongo==3.12.0
pytest==6.2.5
pytest-dotenv==0.5.2
python-dotenv==0.19.1
